+++ test/test_helper.rb
@@ -38,7 +38,7 @@ module ActiveSupport
     def stub_successful_twilio_call(message, user)
       stub_request(:post, "https://api.twilio.com/2010-04-01/Accounts/#{ENV.fetch("TWILIO_ACCOUNT_SID")}/Messages.json")
         .with(
-          body: {"Body" => message, "MessagingServiceSid" => ENV.fetch("TWILIO_MESSAGING_SERVICE_SID"), "StatusCallback" => "/messages/status", "To" => user.phone_number},
+          body: {"Body" => message, "MessagingServiceSid" => ENV.fetch("TWILIO_MESSAGING_SERVICE_SID"), "StatusCallback" => "/messages/twilio_status", "To" => user.phone_number},
         )
         .to_return(status: 200, body: {"body" => message, "status" => "accepted", "sid" => "123"}.to_json)
     end
@@ -46,9 +46,51 @@ module ActiveSupport
     def stub_unsuccessful_twilio_call(message, user)
       stub_request(:post, "https://api.twilio.com/2010-04-01/Accounts/#{ENV.fetch("TWILIO_ACCOUNT_SID")}/Messages.json")
         .with(
-          body: {"Body" => message, "MessagingServiceSid" => ENV.fetch("TWILIO_MESSAGING_SERVICE_SID"), "StatusCallback" => "/messages/status", "To" => user.phone_number},
+          body: {"Body" => message, "MessagingServiceSid" => ENV.fetch("TWILIO_MESSAGING_SERVICE_SID"), "StatusCallback" => "/messages/twilio_status", "To" => user.phone_number},
         )
         .to_return(status: 500, body: {"body" => message, "status" => "failed", "sid" => "123"}.to_json)
+    end
+
+    def stub_successful_aws_call(message, user, aws_message_id: "aws-message-id")
+      stub_request(:post, %r{https://sms-voice\..+\.amazonaws\.com/})
+        .with do |request|
+          body = ::JSON.parse(request.body)
+          body["DestinationPhoneNumber"] == user.phone_number &&
+            body["OriginationIdentity"] == ENV.fetch("AWS_SMS_ORIGINATION_ID") &&
+            body["MessageBody"] == message &&
+            body["MessageType"] == "TRANSACTIONAL"
+        end
+        .to_return(
+          status: 200,
+          body: {"MessageId" => aws_message_id}.to_json,
+          headers: {"Content-Type" => "application/x-amz-json-1.0"},
+        )
+    end
+
+    def stub_unsuccessful_aws_call(message, user)
+      stub_request(:post, %r{https://sms-voice\..+\.amazonaws\.com/})
+        .with do |request|
+          body = ::JSON.parse(request.body)
+          body["DestinationPhoneNumber"] == user.phone_number &&
+            body["MessageBody"] == message
+        end
+        .to_return(
+          status: 400,
+          body: {"__type" => "ValidationException", "Message" => "Invalid phone number"}.to_json,
+          headers: {"Content-Type" => "application/x-amz-json-1.0"},
+        )
+    end
+
+    def stub_sns_verification_success
+      Aws::SNS::MessageVerifier.any_instance.stubs(:authenticate!).returns(true)
+    end
+
+    def sns_notification_envelope(**event_attrs)
+      {
+        "Type" => "Notification",
+        "TopicArn" => "arn:aws:sns:eu-west-2:123:topic",
+        "Message" => event_attrs.transform_keys(&:to_s).to_json,
+      }
     end
 
     def assert_see(text)